requirements.txt
flask>=3.0
opencv-python>=4.8
python-dotenv>=1.0
numpy>=1.24
python-dvr>=0.0.1
